delete citations.go!
